* src/gsad.c (request_handler): Pass gsize to get_system_report_omp.
git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/gsa@6370 423fd1db-d629-0410-8442-d21db03e70f4

+++ src/gsad.c
@@ -2469,7 +2469,7 @@ request_handler (void *cls, struct MHD_Connection *connection,
                     strlen ("/system_report/"))) /* flawfinder: ignore,
                                                     it is a const str */
         {
-          unsigned int res_len;
+          gsize res_len;
           const char *duration;
 
           duration = MHD_lookup_connection_value (connection,
@@ -2485,8 +2485,8 @@ request_handler (void *cls, struct MHD_Connection *connection,
                                        &content_disposition,
                                        &res_len);
           if (res == NULL) return MHD_NO;
-          response = MHD_create_response_from_data (res_len, res,
-                                                    MHD_NO, MHD_YES);
+          response = MHD_create_response_from_data ((unsigned int) res_len,
+                                                    res, MHD_NO, MHD_YES);
           if (content_type != NULL)
             {
               MHD_add_response_header (response, MHD_HTTP_HEADER_CONTENT_TYPE,
